backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
seaborn
langchain
langchain-groq
langchain-experimental
python-multipart
uvicorn[standard]
python-dotenv